Rename package name to Yomitan.
