+++ go.mod
@@ -1,10 +1,10 @@
-module foosoft.net/projects/yomichan-import
+module github.com/themoeway/yomitan-import
 
 go 1.18
 
 require (
-	foosoft.net/projects/jmdict v0.0.0-20220704034116-2e3194aa18d5
-	foosoft.net/projects/zero-epwing-go v0.0.0-20220704035039-bc008453615d
+	github.com/themoeway/jmdict-go v0.0.0-20220704034116-2e3194aa18d5
+	github.com/themoeway/zero-epwing-go v0.0.0-20220704035039-bc008453615d
 	github.com/andlabs/ui v0.0.0-20200610043537-70a69d6ae31e
 	github.com/mattn/go-sqlite3 v1.14.14
 	golang.org/x/exp v0.0.0-20221207211629-99ab8fa1c11f
