+++ jmdict.go
@@ -1,4 +1,4 @@
-package yomichan
+package yomitan
 
 import (
 	"errors"
@@ -7,7 +7,7 @@ import (
 	"strconv"
 	"strings"
 
-	"foosoft.net/projects/jmdict"
+	"github.com/themoeway/jmdict-go"
 	"golang.org/x/exp/slices"
 )
