+++ jmdict_constants.go
@@ -1,4 +1,4 @@
-package yomichan
+package yomitan
 
 type LangCode struct {
 	language string
