+++ jmnedict_generic_terms.go
@@ -1,4 +1,4 @@
-package yomichan
+package yomitan
 
 import (
 	"golang.org/x/exp/slices"
