+++ structured_content.go
@@ -1,4 +1,4 @@
-package yomichan
+package yomitan
 
 type contentAttr struct {
 	lang               string
